+++ webapp/src/components/cards/ObizCard.tsx
@@ -1,4 +1,4 @@
-import { Text, Tag, Center, Link, Box } from "@chakra-ui/react";
+import { Text, Tag, Center, Link, Box, Flex } from "@chakra-ui/react";
 import { OfferIncluded } from "~/server/api/routers/offer";
 import { BarcodeIcon } from "../icons/barcode";
 import NextLink from "next/link";
@@ -6,15 +6,20 @@ import Image from "../ui/Image";
 
 type ObizCardProps = {
   offer: OfferIncluded;
+  fromWidget?: boolean;
 };
 
 export const ObizCard = (props: ObizCardProps) => {
-  const { offer } = props;
+  const { offer, fromWidget } = props;
 
   return (
     <Link
       as={NextLink}
-      href={`/dashboard/offer/${offer.source}/${offer.id}`}
+      href={
+        fromWidget
+          ? `/widget/offer/${offer.id}`
+          : `/dashboard/offer/${offer.source}/${offer.id}`
+      }
       _hover={{ textDecoration: "none" }}
       passHref
     >
@@ -27,26 +32,47 @@ export const ObizCard = (props: ObizCardProps) => {
         gap={2}
         p={4}
       >
-        <Box
-          borderRadius="2.5xl"
+        {offer.image && (
+          <Image
+            src={offer.image.url as string}
+            alt={offer.image.alt as string}
+            width={115}
+            height={70}
+            imageStyle={{
+              width: "115px",
+              height: "70px",
+              transform: "translateY(40%)",
+              marginTop: "-2rem",
+              zIndex: 0,
+            }}
+          />
+        )}
+        <Flex
+          alignItems="center"
+          borderRadius="md"
           border="1px solid"
           borderColor="bgGray"
           p={0.5}
+          bg={"white"}
+          overflow={"hidden"}
+          zIndex={1}
+          height="56px"
+          maxWidth="56px"
         >
           <Image
             src={offer.partner.icon.url as string}
             alt={offer.partner.icon.alt as string}
-            width={40}
-            height={40}
+            width={offer.partner.icon.width || 56}
+            height={offer.partner.icon.height || 56}
             imageStyle={{
-              width: "40px",
-              height: "40px",
-              borderRadius: "20px",
+              width: "56px",
+              maxHeight: "56px",
+              borderRadius: "5px",
               objectFit: "cover",
               objectPosition: "center",
             }}
           />
-        </Box>
+        </Flex>
         <Tag fontWeight={700} w="fit-content" borderRadius="3xl">
           <BarcodeIcon mr={1} /> Bon d'achat
         </Tag>