+++ webapp/src/pages/dashboard/offer/obiz/[id]/index.tsx
@@ -95,12 +95,47 @@ export default function OfferObizPage() {
           mx={4}
         >
           <Flex direction={"column"} mb={2} alignItems={"center"} gap={2}>
-            <Image
-              src={offer.partner.icon.url || ""}
-              width={60}
-              height={offer.partner.icon.height || 50}
-              alt={`Logo ${offer.partner.name}`}
-            />
+            {offer.image && (
+              <Image
+                src={offer.image.url as string}
+                alt={offer.image.alt as string}
+                width={115}
+                height={70}
+                imageStyle={{
+                  width: "115px",
+                  height: "70px",
+                  transform: "translateY(40%)",
+                  marginTop: "-2rem",
+                  zIndex: 0,
+                }}
+              />
+            )}
+            <Flex
+              alignItems="center"
+              borderRadius="md"
+              border="1px solid"
+              borderColor="bgGray"
+              p={0.5}
+              bg={"white"}
+              overflow={"hidden"}
+              zIndex={1}
+              height="56px"
+              maxWidth="56px"
+            >
+              <Image
+                src={offer.partner.icon.url as string}
+                alt={offer.partner.icon.alt as string}
+                width={offer.partner.icon.width || 56}
+                height={offer.partner.icon.height || 56}
+                imageStyle={{
+                  width: "56px",
+                  maxHeight: "56px",
+                  borderRadius: "5px",
+                  objectFit: "cover",
+                  objectPosition: "center",
+                }}
+              />
+            </Flex>
             <Text fontWeight={700} fontSize={"xl"}>
               {offer.partner.name}
             </Text>